[APM] Minor telemetry client refactor
