+++ x-pack/plugins/apm/server/lib/apm_telemetry/collect_data_telemetry/index.ts
@@ -6,40 +6,21 @@
  */
 
 import { merge } from 'lodash';
-import { Logger, SavedObjectsClient } from '@kbn/core/server';
-import { ApmIndicesConfig } from '../../../routes/settings/apm_indices/get_apm_indices';
-import { tasks } from './tasks';
+import { Logger } from '@kbn/core/server';
+import { tasks, TelemetryTaskExecutorParams } from './tasks';
 import { APMDataTelemetry } from '../types';
-import { TelemetryClient } from '../telemetry_client';
 
-type ISavedObjectsClient = Pick<SavedObjectsClient, 'find'>;
-
-export interface TelemetryTaskExecutorParams {
-  telemetryClient: TelemetryClient;
-  indices: ApmIndicesConfig;
-  savedObjectsClient: ISavedObjectsClient;
-}
-
-type TelemetryTaskExecutor = (
-  params: TelemetryTaskExecutorParams
-) => Promise<APMDataTelemetry>;
-
-export interface TelemetryTask {
-  name: string;
-  executor: TelemetryTaskExecutor;
-}
-
-export type CollectTelemetryParams = TelemetryTaskExecutorParams & {
+type CollectTelemetryParams = TelemetryTaskExecutorParams & {
   isProd: boolean;
   logger: Logger;
 };
 
 export function collectDataTelemetry({
   indices,
-  logger,
   telemetryClient,
   savedObjectsClient,
   isProd,
+  logger,
 }: CollectTelemetryParams) {
   return tasks.reduce((prev, task) => {
     return prev.then(async (data) => {
@@ -47,8 +28,8 @@ export function collectDataTelemetry({
       try {
         const time = process.hrtime();
         const next = await task.executor({
-          telemetryClient,
           indices,
+          telemetryClient,
           savedObjectsClient,
         });
         const took = process.hrtime(time);
